+++ cvat/apps/dataset_manager/formats/mots.py
@@ -8,7 +8,7 @@ from datumaro.components.dataset import Dataset
 from datumaro.components.extractor import AnnotationType, ItemTransform
 from pyunpack import Archive
 
-from cvat.apps.dataset_manager.bindings import (CvatTaskDataExtractor,
+from cvat.apps.dataset_manager.bindings import (GetCVATDataExtractor,
     find_dataset_root, match_dm_item)
 from cvat.apps.dataset_manager.util import make_zip_archive
 
@@ -21,9 +21,9 @@ class KeepTracks(ItemTransform):
             if 'track_id' in a.attributes])
 
 @exporter(name='MOTS PNG', ext='ZIP', version='1.0')
-def _export(dst_file, task_data, save_images=False):
-    dataset = Dataset.from_extractors(CvatTaskDataExtractor(
-        task_data, include_images=save_images), env=dm_env)
+def _export(dst_file, instance_data, save_images=False):
+    dataset = Dataset.from_extractors(GetCVATDataExtractor(
+        instance_data, include_images=save_images), env=dm_env)
     dataset.transform(KeepTracks) # can only export tracks
     dataset.transform('polygons_to_masks')
     dataset.transform('boxes_to_masks')