+++ cvat/apps/dataset_manager/formats/vggface2.py
@@ -7,7 +7,7 @@ from tempfile import TemporaryDirectory
 
 from datumaro.components.dataset import Dataset
 
-from cvat.apps.dataset_manager.bindings import CvatTaskDataExtractor, \
+from cvat.apps.dataset_manager.bindings import GetCVATDataExtractor, \
     import_dm_annotations
 from cvat.apps.dataset_manager.util import make_zip_archive
 
@@ -15,19 +15,19 @@ from .registry import dm_env, exporter, importer
 
 
 @exporter(name='VGGFace2', ext='ZIP', version='1.0')
-def _export(dst_file, task_data, save_images=False):
-    dataset = Dataset.from_extractors(CvatTaskDataExtractor(
-        task_data, include_images=save_images), env=dm_env)
+def _export(dst_file, instance_data, save_images=False):
+    dataset = Dataset.from_extractors(GetCVATDataExtractor(
+        instance_data, include_images=save_images), env=dm_env)
     with TemporaryDirectory() as temp_dir:
         dataset.export(temp_dir, 'vgg_face2', save_images=save_images)
 
         make_zip_archive(temp_dir, dst_file)
 
 @importer(name='VGGFace2', ext='ZIP', version='1.0')
-def _import(src_file, task_data):
+def _import(src_file, instance_data):
     with TemporaryDirectory() as tmp_dir:
         zipfile.ZipFile(src_file).extractall(tmp_dir)
 
         dataset = Dataset.import_from(tmp_dir, 'vgg_face2', env=dm_env)
         dataset.transform('rename', r"|([^/]+/)?(.+)|\2|")
-        import_dm_annotations(dataset, task_data)
+        import_dm_annotations(dataset, instance_data)