+++ tests/cypress/integration/actions_projects_models/case_94_move_task_between_projects.js
@@ -11,16 +11,16 @@ context('Move a task between projects.', () => {
         label: 'car',
         attrName: 'color',
         attrVaue: 'red',
-        multiAttrParams: false
-    }
+        multiAttrParams: false,
+    };
 
     const secondProject = {
         name: `Second project case ${caseID}`,
         label: 'bicycle',
         attrName: 'color',
         attrVaue: 'yellow',
-        multiAttrParams: false
-    }
+        multiAttrParams: false,
+    };
 
     const taskName = `Task case ${caseID}`;
     const imagesCount = 1;
@@ -39,14 +39,24 @@ context('Move a task between projects.', () => {
     const attachToProject = false;
     const multiAttrParams = false;
 
-    function checkTask (project, expectedResult) {
+    function checkTask(project, expectedResult) {
         cy.goToProjectsList();
         cy.openProject(project);
         cy.get('.cvat-tasks-list-item').should(expectedResult);
     }
 
     before(() => {
-        cy.imageGenerator(imagesFolder, imageFileName, width, height, color, posX, posY, firtsProject.label, imagesCount);
+        cy.imageGenerator(
+            imagesFolder,
+            imageFileName,
+            width,
+            height,
+            color,
+            posX,
+            posY,
+            firtsProject.label,
+            imagesCount,
+        );
         cy.createZipArchive(directoryToArchive, archivePath);
         cy.visit('/');
         cy.login();
@@ -54,8 +64,20 @@ context('Move a task between projects.', () => {
 
     beforeEach(() => {
         cy.goToProjectsList();
-        cy.createProjects(firtsProject.name, firtsProject.label, firtsProject.attrName, firtsProject.attrVaue, firtsProject.multiAttrParams);
-        cy.createProjects(secondProject.name, secondProject.label, secondProject.attrName, secondProject.attrVaue, secondProject.multiAttrParams);
+        cy.createProjects(
+            firtsProject.name,
+            firtsProject.label,
+            firtsProject.attrName,
+            firtsProject.attrVaue,
+            firtsProject.multiAttrParams,
+        );
+        cy.createProjects(
+            secondProject.name,
+            secondProject.label,
+            secondProject.attrName,
+            secondProject.attrVaue,
+            secondProject.multiAttrParams,
+        );
         cy.openProject(firtsProject.name);
         cy.createAnnotationTask(
             taskName,